Added Co8 dialog scripts.

+++ ScriptConversion/ScriptConversion.cs
@@ -3,13 +3,18 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Text;
+using System.Text.RegularExpressions;
 using IronPython.Modules;
 using IronPython.Runtime.Operations;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Formatting;
 using Microsoft.CodeAnalysis.Text;
 using SpicyTemple.Core.IO.TroikaArchives;
+using SpicyTemple.Core.Systems;
+using SpicyTemple.Core.Systems.Dialog;
 using SpicyTemple.Core.TigSubsystems;
 
 namespace ScriptConversion
@@ -22,10 +27,24 @@ namespace ScriptConversion
 
         private readonly ScriptConverter converter;
 
+        private readonly List<PythonScript> _convertedScripts = new List<PythonScript>();
+
         // Format the script code before writing it to a file
         private AdhocWorkspace workspace;
 
         private Project project;
+
+        private bool _writeScripts = false;
+
+        private bool _writeDialog = true;
+
+        private static readonly Dictionary<string, GuessedType> DialogContext =
+            new Dictionary<string, GuessedType>
+            {
+                {"picked_line", GuessedType.Integer},
+                {"pc", GuessedType.Object},
+                {"npc", GuessedType.Object}
+            };
 
         public ScriptConversion(string outputDirectory)
         {
@@ -59,6 +78,283 @@ namespace ScriptConversion
             var conversion = new ScriptConversion(outputDir);
             conversion.LoadTyping(outputDir);
             conversion.ConvertScripts();
+            conversion.ConvertDialog();
+        }
+
+        private void ConvertDialog()
+        {
+            var dialogFiles = Tig.FS.ListDirectory("dlg")
+                .Where(f => f.EndsWith(".dlg"))
+                .Distinct()
+                .ToList();
+
+            Console.WriteLine($"Found {dialogFiles.Count} dialog files.");
+
+            var dlgFilePattern = new Regex(@"(\d{5}).*");
+
+            foreach (var dialogFile in dialogFiles)
+            {
+                var m = dlgFilePattern.Match(dialogFile);
+                if (!m.Success)
+                {
+                    Console.WriteLine($"Skipping dialog file that doesn't match expected pattern: {dialogFile}'.");
+                    continue;
+                }
+
+                var scriptId = int.Parse(m.Groups[1].Value);
+                var associatedScript = _convertedScripts.FirstOrDefault(s => s.Type == ScriptType.Object
+                                                                             && s.ScriptId == scriptId);
+                if (associatedScript == null)
+                {
+                    Console.WriteLine($"Dialog file {dialogFile} with id {scriptId} has no associated script!");
+                    continue;
+                }
+
+                var outputDir = Path.Join(
+                    Path.GetDirectoryName(associatedScript.OutputPath),
+                    "Dialog"
+                );
+                Directory.CreateDirectory(Path.Combine(_outputDirectory, outputDir));
+
+                var outputPath = Path.Join(outputDir,
+                    Path.GetFileNameWithoutExtension(associatedScript.OutputPath) + "Dialog.cs");
+                if (_writeDialog)
+                {
+                    File.Delete(Path.Combine(_outputDirectory, outputPath));
+                }
+
+                var dialogContent = Tig.FS.ReadTextFile("dlg/" + dialogFile);
+                var parser = new DialogScriptParser(dialogFile, dialogContent);
+
+                // We're going to build a class file for the dialog script that contains two methods,
+                // and extends from the object script of the same script id
+
+                var conditions = new List<(int, string, string)>();
+                var effects = new List<(int, string, string)>();
+                var skillChecks = new List<(int, SkillCheck)>();
+
+                while (parser.GetSingleLine(out var dialogLine, out var fileLine))
+                {
+                    var effectPython = dialogLine.effectField;
+                    if (effectPython == "pc.barter(npc)" &&
+                        dialogLine.txt.StartsWith("b:", StringComparison.OrdinalIgnoreCase))
+                    {
+                        continue; // Skip bogus barter effect (this is implied by B:)
+                    }
+
+                    try
+                    {
+                        if (!string.IsNullOrEmpty(effectPython))
+                        {
+                            var converted = converter.ConvertSnippet(effectPython, associatedScript, DialogContext);
+                            effects.Add((dialogLine.key, effectPython, converted));
+                        }
+                    }
+                    catch (Exception e)
+                    {
+                        Console.WriteLine(
+                            $"[e] Failed to convert effect for line {dialogLine.key} ({effectPython}): {e}");
+                        effects.Add((dialogLine.key, effectPython, null));
+                    }
+
+                    if (dialogLine.IsPcLine)
+                    {
+                        var conditionPython = dialogLine.testField;
+                        try
+                        {
+                            if (!string.IsNullOrEmpty(conditionPython))
+                            {
+                                foreach (var skillCheck in converter.FindSkillChecks(conditionPython))
+                                {
+                                    skillChecks.Add((dialogLine.key, skillCheck));
+                                }
+
+                                var converted =
+                                    converter.ConvertSnippet(conditionPython, associatedScript, DialogContext);
+                                conditions.Add((dialogLine.key, conditionPython, converted));
+                            }
+                        }
+                        catch (Exception e)
+                        {
+                            Console.WriteLine(
+                                $"[e] Failed to convert condition for line {dialogLine.key} ({conditionPython}): {e}");
+                            conditions.Add((dialogLine.key, conditionPython, null));
+                        }
+                    }
+                }
+
+                if (_writeDialog)
+                {
+                    var dialogScript = new StringBuilder();
+
+                    dialogScript.AppendLine(@"
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using SpicyTemple.Core.GameObject;
+using SpicyTemple.Core.Systems;
+using SpicyTemple.Core.Systems.Dialog;
+using SpicyTemple.Core.Systems.Feats;
+using SpicyTemple.Core.Systems.D20;
+using SpicyTemple.Core.Systems.Script;
+using SpicyTemple.Core.Systems.Spells;
+using SpicyTemple.Core.Systems.GameObjects;
+using SpicyTemple.Core.Systems.D20.Conditions;
+using SpicyTemple.Core.Location;
+using SpicyTemple.Core.Systems.ObjScript;
+using SpicyTemple.Core.Ui;
+using System.Linq;
+using SpicyTemple.Core.Systems.Script.Extensions;
+using SpicyTemple.Core.Utils;
+using static SpicyTemple.Core.Systems.Script.ScriptUtilities;
+");
+                    dialogScript.AppendLine("namespace " + associatedScript.Namespace + ".Dialog");
+                    dialogScript.AppendLine("{");
+                    dialogScript.Append("[DialogScript(").Append(associatedScript.ScriptId).AppendLine(")]");
+                    dialogScript.AppendLine("public class " + associatedScript.ClassName + "Dialog : " +
+                                            associatedScript.ClassName + ", IDialogScript");
+                    dialogScript.AppendLine("{");
+
+                    WriteDialogMethod(dialogScript, conditions, false);
+                    WriteDialogMethod(dialogScript, effects, true);
+                    WriteSkillChecksMethod(dialogScript, skillChecks);
+
+                    dialogScript.AppendLine("}");
+                    dialogScript.AppendLine("}");
+
+                    var parseOptions = new CSharpParseOptions(LanguageVersion.Latest);
+                    var syntaxTree = CSharpSyntaxTree.ParseText(dialogScript.ToString(), parseOptions);
+                    var formattedNode = Formatter.Format(syntaxTree.GetRoot(), workspace);
+
+                    File.WriteAllText(Path.Join(_outputDirectory, outputPath), formattedNode.ToFullString());
+                }
+            }
+        }
+
+        private static void WriteDialogMethod(StringBuilder dialogScript, List<(int, string, string)> conditions,
+            bool effects)
+        {
+            if (effects)
+            {
+                dialogScript.AppendLine(
+                    "public void ApplySideEffect(GameObjectBody npc, GameObjectBody pc, int lineNumber, string originalScript)");
+            }
+            else
+            {
+                dialogScript.AppendLine(
+                    "public bool CheckPrecondition(GameObjectBody npc, GameObjectBody pc, int lineNumber, string originalScript)");
+            }
+
+            dialogScript.AppendLine("{");
+            dialogScript.AppendLine("switch (lineNumber)");
+            dialogScript.AppendLine("{");
+            for (var i = 0; i < conditions.Count; i++)
+            {
+                if (conditions[i].Item1 == -1)
+                {
+                    continue; // Already printed elsewhere
+                }
+
+                dialogScript.Append("case ").Append(conditions[i].Item1).AppendLine(":");
+                // Collapse identical conditions
+                for (var j = i + 1; j < conditions.Count; j++)
+                {
+                    if (conditions[j].Item2 == conditions[i].Item2)
+                    {
+                        dialogScript.Append("case ").Append(conditions[j].Item1).AppendLine(":");
+                        conditions[j] = (-1, null, null);
+                    }
+                }
+
+                var escapedOriginal = SyntaxFactory
+                    .LiteralExpression(SyntaxKind.StringLiteralExpression, SyntaxFactory.Literal(conditions[i].Item2))
+                    .ToFullString();
+                dialogScript.Append("Trace.Assert(originalScript == ").Append(escapedOriginal).AppendLine(");");
+                var converted = conditions[i].Item3;
+                if (converted != null)
+                {
+                    if (!effects)
+                    {
+                        dialogScript.Append("return ");
+                    }
+
+                    dialogScript.Append(converted);
+                    if (converted.EndsWith(";"))
+                    {
+                        dialogScript.AppendLine();
+                    }
+                    else
+                    {
+                        dialogScript.AppendLine(";");
+                    }
+
+                    if (effects)
+                    {
+                        dialogScript.AppendLine("break;");
+                    }
+                }
+                else
+                {
+                    dialogScript.AppendLine("throw new NotSupportedException(\"Conversion failed.\");");
+                }
+            }
+
+            dialogScript.AppendLine("default:");
+            dialogScript.AppendLine("Trace.Assert(originalScript == null);");
+            if (effects)
+            {
+                dialogScript.AppendLine("return;");
+            }
+            else
+            {
+                dialogScript.AppendLine("return true;");
+            }
+
+            dialogScript.AppendLine("}");
+            dialogScript.AppendLine("}");
+        }
+
+        private void WriteSkillChecksMethod(StringBuilder dialogScript, List<(int, SkillCheck)> skillChecks)
+        {
+            dialogScript.AppendLine("public bool TryGetSkillCheck(int lineNumber, out DialogSkillCheck skillCheck)");
+            dialogScript.AppendLine("{");
+            dialogScript.AppendLine("switch (lineNumber)");
+            dialogScript.AppendLine("{");
+
+            for (var i = 0; i < skillChecks.Count; i++)
+            {
+                var (lineNumber, skillCheck) = skillChecks[i];
+                if (lineNumber == -1)
+                {
+                    continue; // already handled elsewhere
+                }
+
+                dialogScript.Append("case ").Append(lineNumber).AppendLine(":");
+                // Collapse identical skill checks
+                for (var j = i + 1; j < skillChecks.Count; j++)
+                {
+                    if (skillCheck.Skill == skillChecks[j].Item2.Skill
+                        && skillCheck.Ranks == skillChecks[j].Item2.Ranks)
+                    {
+                        dialogScript.Append("case ").Append(skillChecks[j].Item1).AppendLine(":");
+                        skillChecks[j] = (-1, default);
+                    }
+                }
+
+                dialogScript.Append("skillCheck = new DialogSkillCheck(");
+                dialogScript.Append("SkillId.");
+                dialogScript.Append(skillCheck.Skill.ToString());
+                dialogScript.Append(", ");
+                dialogScript.Append(skillCheck.Ranks);
+                dialogScript.AppendLine(");");
+                dialogScript.AppendLine("return true;");
+            }
+
+            dialogScript.AppendLine("default:");
+            dialogScript.AppendLine("skillCheck = default;");
+            dialogScript.AppendLine("return false;");
+            dialogScript.AppendLine("}");
+            dialogScript.AppendLine("}");
         }
 
         private void LoadTyping(string dir)
@@ -116,7 +412,8 @@ namespace ScriptConversion
                 }
 
                 var suffix = 1;
-                Console.WriteLine("Scripts with clashing names: " + string.Join(", ", grouping.Select(c => c.Filename)));
+                Console.WriteLine("Scripts with clashing names: " +
+                                  string.Join(", ", grouping.Select(c => c.Filename)));
                 foreach (var script in grouping)
                 {
                     script.ClassName += suffix;
@@ -134,7 +431,10 @@ namespace ScriptConversion
                     Directory.CreateDirectory(parentDir);
                 }
 
-                File.Delete(outputFile);
+                if (_writeScripts)
+                {
+                    File.Delete(outputFile);
+                }
             }
 
             Console.WriteLine($"Parsed {scripts.Count} scripts...");
@@ -152,6 +452,8 @@ namespace ScriptConversion
                     var script = remaining[i];
                     if (!HasUnmetDependencies(script, modules))
                     {
+                        _convertedScripts.Add(script);
+
                         try
                         {
                             ConvertScript(script, modules);
@@ -218,10 +520,13 @@ namespace ScriptConversion
 
             var document = workspace.AddDocument(project.Id, script.OutputPath, SourceText.From(convertedSource));
 
-            var syntaxTree = CSharpSyntaxTree.ParseText(document.GetTextAsync().Result);
-            var formattedNode = Formatter.Format(syntaxTree.GetRoot(), workspace);
+            if (_writeScripts)
+            {
+                var syntaxTree = CSharpSyntaxTree.ParseText(document.GetTextAsync().Result);
+                var formattedNode = Formatter.Format(syntaxTree.GetRoot(), workspace);
 
-            File.WriteAllText(outputFile, formattedNode.ToFullString());
+                File.WriteAllText(outputFile, formattedNode.ToFullString());
+            }
         }
     }
 }
